Optimize findPrimesInRange for time
Before mutex was being used as a means to keep the threads from writing and accessing the vector storing primes at the same time. However this caused significant slowdown as each thread would pause the others to write to the vector.

Instead now we have a vector of vectors, one for each thread, that they can write to concurrently.

Significant time saver.